exercise 2.14 finally complete added a confirmation for delete which i forgot to add

+++ part2/Phonebook/src/App.jsx
@@ -48,9 +48,13 @@ const App = () => {
     setFilter(event.target.value)
   }
   const handleDelete = (id) => {
-    personsService.deletePerson(id).then(() => {
-      setPersons(persons.filter((person)=>person.id!==id))
-    })
+    if (window.confirm(`Delete ${persons.find(person=>person.id===id).name}?`)) {
+      personsService.deletePerson(id).then(() => {
+        setPersons(
+          persons.filter(person => person.id !== id)
+        );
+      });
+    }
   }
   return (
     <div>
